+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Added simplified output image resolution picker.
 - Added upload and management of input/mask files inside of projects.
 
 ## Improvements